chore(main): release 1.2.8
Enhancements:
Improved the flexibility of the archive function, allowing the user to archive any note, not just chat notes. 

The status message of a note being moved to the archive is no longer appended to the note body itself. Originally, it seemed like a good way to show the user that the note they were looking at had moved, but in retrospect I think that modifying the notes content should not be done. There is a brief status message displayed in the top right corner, and the file browser is updated accordingly. I think after the first use of the feature the user will understand what happened. 

Bugfixes:

Finally, updated a few settings options to render in the correct position.